+++ Sources/TableViewDriver.swift
@@ -300,11 +300,15 @@ extension TableViewDriver: UITableViewDelegate {
     }
 
     /// :nodoc:
-    public func tableView(_ tableView: UITableView, editActionsForRowAt indexPath: IndexPath) -> [UITableViewRowAction]? {
-        if let cellViewModel = self.tableViewModel?[ifExists: indexPath] as? TableViewCellModelEditActions {
-            return cellViewModel.editActions
-        }
-        return nil
+    public func tableView(_ tableView: UITableView, leadingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
+        guard let editActions = self.tableViewModel?[ifExists: indexPath] as? TableViewCellModelEditActions else { return nil }
+        return editActions.leadingSwipeActionConfiguration
+    }
+
+    /// :nodoc:
+    public func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
+        guard let editActions = self.tableViewModel?[ifExists: indexPath] as? TableViewCellModelEditActions else { return nil }
+        return editActions.trailingSwipeActionConfiguration
     }
 
     /// :nodoc: